refactor(webos-common): rename generateWriter to generateModeWriter
Rename generateWriter.js → generateModeWriter.js and its exported
writeGenResources() → writeGenerateModeResources() for clarity, matching
the pseudoWriter module naming.
