+++ packages/ilib-loctool-webos-dart/DartFileType.js
@@ -24,12 +24,11 @@ var Locale = require("ilib/lib/Locale.js");
 var DartFile = require("./DartFile.js");
 var JsonResourceFileType = require("ilib-loctool-webos-json-resource");
 var Utils = require("loctool/lib/utils.js")
-var { utils: pluginUtils, translationResolver, pseudoWriter, generateWriter } = require("ilib-loctool-webos-common");
+var { utils: pluginUtils, translationResolver, pseudoWriter, generateModeWriter } = require("ilib-loctool-webos-common");
 var buildResolver = translationResolver.buildResolver;
 var resolveTranslation = translationResolver.resolveTranslation;
-var filterGenResources = generateWriter.filterGenResources;
 var writePseudoResources = pseudoWriter.writePseudoResources;
-var writeGenResources = generateWriter.writeGenResources;
+var writeGenerateModeResources = generateModeWriter.writeGenerateModeResources;
 
 var DartFileType = function(project) {
     this.type = "x-dart";
@@ -181,9 +180,9 @@ DartFileType.prototype.write = function(translations, locales) {
         }.bind(this));
 
     // Build resolver: detects common project data and prepares policy-based lookup.
-    var resolver = buildResolver(db, translations, this.project.getProjectId(), { includeUniversal: true });
-
     if (mode === "localize") {
+        // Build resolver: detects common project data and prepares policy-based lookup.
+        var resolver = buildResolver(db, translations, this.project.getProjectId(), { includeUniversal: true });
         for (var i = 0; i < resources.length; i++) {
             res = resources[i];
             // for each extracted string, write out the translations of it
@@ -221,20 +220,16 @@ DartFileType.prototype.write = function(translations, locales) {
         });
     } else {
         // generate mode
-        this.genresources = filterGenResources(
-            this.project.getTranslations(translationLocales),
-            this.project.getProjectId(),
-            this.datatype,
-            { includeUniversal: true }
-        );
-        writeGenResources({
+        writeGenerateModeResources({
             project: this.project,
             translationLocales: translationLocales,
-            genresources: this.genresources,
+            resources: this.project.getTranslations(translationLocales),
+            selfDatatype: this.datatype,
             resFileType: resFileType,
             db: db,
             deviceType: deviceType,
-            dedupByBaseTranslation: false
+            dedupByBaseTranslation: false,
+            includeUniversal: true
         });
     }
 };
